page: Exclude the "service" page from the website
Currently, the "service" page is empty and displayed. It doesn't make
much sense to show an empty page like this.

Add "eleventyExcludeFromCollections: true" in the "service.md" file to
avoid this page being displayed, but don't delete this file for now. In
the future, if we plan to use this file, just remove the exclusion
attribute.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>