body.php: Add the `[lang]` attribute to the `<html>` element
If the page does not specify a language attribute, the screen reader
assumes that the page is in the default language that the user selected
when setting up the screen reader. This will be a problem for some
people and will have difficulty pronouncing the text on the page
correctly.

Signed-off-by: Memet Zx <memetpea25@gmail.com>
Link: https://lore.gnuweeb.org/gwml/20230129114727.33428-3-zxce3@gnuweeb.org
[ammarfaizi2: Fix commit message word-wrapping]
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>