head.php: Fix issue with css/js head section
When visiting a web page that ends with a forward slash `/`,
the associated CSS and JS files may fail to load properly.
This issue can be resolved by adding the `BASE_URL` to the links used to
call the CSS and JS files.
By doing so, the browser will be able to locate and load the required
resources, ensuring that the web page displays correctly.

Signed-off-by: Memet Zx <zxce3@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20230415144805.282728-2-zxce3@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>