README.md: Initialize the information
The README.md file provides important information about the project,
including its purpose, how to install and run it, and any dependencies
or requirements.
This file will help other developers understand the project and
contribute to it more easily.
Additionally, it will serve as a reference for future development and
maintenance tasks.

Signed-off-by: Memet Zx <zxce3@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20230415144805.282728-3-zxce3@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>