feat(page): improve UX for copying text
This commit introduces an improvement to the user experience when
copying text. Now, when a user clicks the copy button or the copy text
area, the background color of the copy button will change. This visual
feedback enhances the user experience by providing a highlight effect
upon clicking, making it more intuitive and user-friendly.

Signed-off-by: Muhammad Rizki <kiizuha@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20230924094532.1662-5-kiizuha@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>